Move Bytes from utils to schemas
