+++ griptape/schemas/bytes_field.py
