--- griptape/utils/marshmallow/fields/__init__.py
@@ -1,4 +1,0 @@
-from .bytes import Bytes
-
-
-__all__ = ["Bytes"]
