+++ griptape/utils/stream.py
@@ -7,8 +7,8 @@ from griptape.events.completion_chunk_event import CompletionChunkEvent
 from griptape.events.event_listener import EventListener
 from griptape.events.base_event import BaseEvent
 from griptape.events.finish_structure_run_event import FinishStructureRunEvent
+from griptape.events.finish_prompt_event import FinishPromptEvent
 from attrs import field, define, Factory
-from griptape.events import BaseEvent, FinishStructureRunEvent, FinishPromptEvent
 
 if TYPE_CHECKING:
     from griptape.structures import Structure
